requirements.txt
numpy==1.18.1
Pillow==7.0.0
pkg-resources==0.0.0
scipy==1.4.1
